refact: 아낀 항목 패키지 구조 변경
