+++ src/main/java/com/akkin/gulbi/create/GulbiCreateController.java
@@ -1,0 +1,28 @@
+package com.akkin.gulbi.create;
+
+import com.akkin.auth.AuthRequired;
+import com.akkin.gulbi.create.dto.GulbiCreateForm;
+import com.akkin.gulbi.create.dto.GulbiCreateResponse;
+import com.akkin.login.dto.AuthMember;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import javax.servlet.http.HttpServletRequest;
+
+@RequiredArgsConstructor
+@RequestMapping("/api/gulbis")
+@RestController
+public class GulbiCreateController {
+
+    private final GulbiCreateService gulbiCreateService;
+
+    @AuthRequired
+    @PostMapping
+    public GulbiCreateResponse saveGulbi(@RequestBody GulbiCreateForm form, HttpServletRequest request) {
+        AuthMember authMember = (AuthMember) request.getAttribute("authMember");
+        return gulbiCreateService.createGulbi(authMember.getId(), form);
+    }
+}
