+++ src/main/java/com/akkin/gulbi/create/dto/GulbiCreateForm.java
@@ -1,14 +1,18 @@
-package com.akkin.gulbi.dto;
+package com.akkin.gulbi.create.dto;
 
 import com.akkin.gulbi.Category;
 import com.akkin.gulbi.Gulbi;
 import com.akkin.member.Member;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Setter
 @Getter
-public class GulbiUpdateForm {
+@AllArgsConstructor
+@NoArgsConstructor
+public class GulbiCreateForm {
 
     private Integer year;
 
@@ -28,15 +32,16 @@ public class GulbiUpdateForm {
 
     public Gulbi dtoToEntity(Member member) {
         return Gulbi.builder()
-                .member(member)
-                .year(year)
-                .month(month)
-                .day(day)
-                .category(category)
-                .content(content)
-                .how(how)
-                .expectCost(expectCost)
-                .realCost(realCost)
-                .build();
+            .member(member)
+            .year(year)
+            .month(month)
+            .day(day)
+            .category(category)
+            .content(content)
+            .how(how)
+            .expectCost(expectCost)
+            .realCost(realCost)
+            .build();
     }
+
 }
