--- src/main/java/com/akkin/gulbi/dto/GulbiCreateResponses.java
@@ -1,16 +1,0 @@
-package com.akkin.gulbi.dto;
-
-import java.util.List;
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-public class GulbiCreateResponses {
-
-    List<GulbiCreateResponse> entries;
-
-    public GulbiCreateResponses(List<GulbiCreateResponse> entries) {
-        this.entries = entries;
-    }
-}
