+++ src/main/java/com/akkin/gulbi/read/dto/GulbiReadResponse.java
@@ -1,13 +1,13 @@
-package com.akkin.gulbi.dto;
+package com.akkin.gulbi.read.dto;
 
 import com.akkin.gulbi.Category;
 import com.akkin.gulbi.Gulbi;
 import lombok.Getter;
-import lombok.Setter;
+import lombok.NoArgsConstructor;
 
-@Setter
 @Getter
-public class GulbiUpdateResponse {
+@NoArgsConstructor
+public class GulbiReadResponse {
 
     private Long id;
 
@@ -27,7 +27,7 @@ public class GulbiUpdateResponse {
 
     private Integer realCost;
 
-    public GulbiUpdateResponse(Gulbi gulbi) {
+    public GulbiReadResponse(Gulbi gulbi) {
         this.id = gulbi.getId();
         this.year = gulbi.getSaveYear();
         this.month = gulbi.getSaveMonth();
