+++ src/main/java/com/akkin/gulbi/update/dto/GulbiUpdateForm.java
@@ -1,18 +1,14 @@
-package com.akkin.gulbi.dto;
+package com.akkin.gulbi.update.dto;
 
 import com.akkin.gulbi.Category;
 import com.akkin.gulbi.Gulbi;
 import com.akkin.member.Member;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Setter
 @Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class GulbiCreateForm {
+public class GulbiUpdateForm {
 
     private Integer year;
 
@@ -32,16 +28,15 @@ public class GulbiCreateForm {
 
     public Gulbi dtoToEntity(Member member) {
         return Gulbi.builder()
-            .member(member)
-            .year(year)
-            .month(month)
-            .day(day)
-            .category(category)
-            .content(content)
-            .how(how)
-            .expectCost(expectCost)
-            .realCost(realCost)
-            .build();
+                .member(member)
+                .year(year)
+                .month(month)
+                .day(day)
+                .category(category)
+                .content(content)
+                .how(how)
+                .expectCost(expectCost)
+                .realCost(realCost)
+                .build();
     }
-
 }
